+++ CHANGES.md
@@ -2,5 +2,7 @@
 
 * Add support for https pen
 * Add setting to select which embed version to use and set version 2 as default value
+* Add setting to select which tabs must be displayed by default
 * Add a theme selector setting with CodePen default themes
+* Add settings to override theme
 * Add a preview panel